fix(sql): revert migration syntax due to wrong gen type

+++ db/migration/000003_alter_users_role_column.up.sql
@@ -1,5 +1,10 @@
-ALTER TABLE users
-    DROP COLUMN IF EXISTS role;
+DO $$
+BEGIN
+    IF EXISTS(SELECT 1 FROM information_schema.tables WHERE table_name = 'users') THEN
+        ALTER TABLE users
+            DROP COLUMN IF EXISTS role;
 
-ALTER TABLE users
-    ADD COLUMN IF NOT EXISTS role access NOT NULL;
+        ALTER TABLE users
+            ADD COLUMN IF NOT EXISTS role access NOT NULL;
+    END IF;
+END $$;
